HBT-180

Approved-by: Mohammad Nurul Islam Shihan

+++ src/main/java/com/improvedigital/prebid/server/auction/model/Floor.java
@@ -3,11 +3,13 @@ package com.improvedigital.prebid.server.auction.model;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Value;
 
+import java.math.BigDecimal;
+
 @Value(staticConstructor = "of")
 public class Floor {
 
     @JsonProperty("bidFloor")
-    double bidFloor;
+    BigDecimal bidFloor;
 
     @JsonProperty("bidFloorCur")
     String bidFloorCur;
